core/requirements.txt
Django==5.0.2
psycopg2-binary==2.9.9
djangorestframework==3.14.0
python-dotenv==1.0.1
drf-yasg==1.21.7 
django_cors_headers==4.3.*
